requirements.txt
torch
torchvision
transformers
Pillow
opencv-python
matplotlib
numpy
streamlit
